Closer to OP's idea
I changed the CSS a bit, made it follow the mouse, but there's a problem where if you hover back over the title link it creates a small box. Basically doubling up the divs... I'm not comfortable with JSON or whatever is going on here, but I consider myself a tinkerer so I hope it's in the right direction.